feat: add shared publish_event helper
